Made sample hero creation more granular
Instead of performing all hero creation steps at a stroke, Create_Shikari and its sister methods (and helper methods in SampleMaker) now take a step argument and perform steps up to and including that number. This allows the tester to zip through as much or as little of the process as needed before manually performing a particular step that needs inspection.