Rewrote behavior for "add 1 or more secondary Powers/Qualities" in AddArchetype
Once the initial secondary Power/Quality has been added, "assign as many dice as you want to [secondary list] and the rest to [tertiary list]" becomes functionally identical to "add the rest of your dice to [secondary list + tertiary list]", which requires fewer arbitrary choices. Changed AddArchetype to follow the simpler pattern, and the sample hero inputs to work with the new code.
Adjusted the width of SelectWindow in ChoosePQDieSize to handle longer names like Lightning Calculator
Allowed ChoosePQ to extend the length of a section of the Power/Quality list to cover the whole thing IF the prompt is being presented as a SelectWindow instead of text
Fixed entry options appearing twice per entry when the user is prompted to choose a section of a Power/Quality list