Progress on ModeFrame.Resize
I did some science to it and saved the results
Should I be committing this to a branch that's gonna get used as-is? No! But this isn't that branch, and working with the difference between config(width=) and winfo_width() for so long is starting to convince me I should just use wraplength.